refactor: notice api typescript
